+++ openbis_ng_ui/srcTest/js/components/tools/browser/ToolBrowserComponentOpenClose.test.js
@@ -46,7 +46,8 @@ async function testOpenClose() {
       { level: 0, text: 'Entity Validation Plugins' },
       { level: 0, text: 'Queries' },
       { level: 0, text: 'History' },
-      { level: 0, text: 'Import' }
+      { level: 0, text: 'Import' },
+      { level: 0, text: 'Active Users Report' }
     ]
   })
 
@@ -65,7 +66,8 @@ async function testOpenClose() {
       { level: 1, text: testEntityValidationJythonPlugin.name },
       { level: 0, text: 'Queries' },
       { level: 0, text: 'History' },
-      { level: 0, text: 'Import' }
+      { level: 0, text: 'Import' },
+      { level: 0, text: 'Active Users Report' }
     ]
   })
 
@@ -85,7 +87,8 @@ async function testOpenClose() {
       { level: 0, text: 'Queries' },
       { level: 1, text: testQuery.name },
       { level: 0, text: 'History' },
-      { level: 0, text: 'Import' }
+      { level: 0, text: 'Import' },
+      { level: 0, text: 'Active Users Report' }
     ]
   })
 
@@ -104,7 +107,8 @@ async function testOpenClose() {
       { level: 1, text: testEntityValidationJythonPlugin.name },
       { level: 0, text: 'Queries' },
       { level: 0, text: 'History' },
-      { level: 0, text: 'Import' }
+      { level: 0, text: 'Import' },
+      { level: 0, text: 'Active Users Report' }
     ]
   })
 
@@ -121,7 +125,8 @@ async function testOpenClose() {
       { level: 1, text: testEntityValidationJythonPlugin.name },
       { level: 0, text: 'Queries' },
       { level: 0, text: 'History' },
-      { level: 0, text: 'Import' }
+      { level: 0, text: 'Import' },
+      { level: 0, text: 'Active Users Report' }
     ]
   })
 
@@ -137,7 +142,8 @@ async function testOpenClose() {
       { level: 0, text: 'Entity Validation Plugins' },
       { level: 0, text: 'Queries' },
       { level: 0, text: 'History' },
-      { level: 0, text: 'Import' }
+      { level: 0, text: 'Import' },
+      { level: 0, text: 'Active Users Report' }
     ]
   })
 }